Bug 1250356: rollup of changes for webrtc after applying webrtc.org v49 update r=pkerr,ng,pehrsons,etc See ssh://hg.mozilla.org/users/paulrkerr_gmail.com/webrtc49_merge/ for the patch development history.

+++ media/webrtc/trunk/webrtc/modules/audio_processing/aec/aec_resampler.c
@@ -102,7 +102,7 @@ void WebRtcAec_ResampleLinear(void* resampInst,
     mm++;
 
     tnew = be * mm + obj->position;
-    tn = (int)tnew;
+    tn = (size_t)tnew;
   }
 
   *size_out = mm;
